Update intf.h
Add INTF_FLAG_RUNNING.

On Linux, INTF_FLAG_UP does not mean the interface is functional.  For example, an ethernet interface with no cable connected may have the UP flag set.  The UP flag on Linux is essentially an administrative state, not an operational state.  Hence it's often useful to get the RUNNING bit.  Allow this with dnet instead of having to issue ioctls outside of dnet just to get the RUNNING bit.